VQA-model-service/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
copier==6.2.0
dnspython==2.7.0
dunamai==1.23.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
iteration_utilities==0.13.0
Jinja2==3.1.4
jinja2-ansible-filters==1.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
plumbum==1.9.0
prompt-toolkit==3.0.36
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
pyyaml-include==2.2
questionary==2.0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
setuptools==75.6.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
